useExtensionsSection
